add url protocol when it is missing + show toast when user add an invalid url
