+++ components/JournalyEditor/helpers.ts
@@ -1,5 +1,26 @@
 import isUrl from 'is-url'
 import { Editor, Transforms, Element as SlateElement, Range } from 'slate'
+import { toast } from 'react-toastify'
+import { TFunction } from 'next-i18next'
+
+export type ButtonType = 'mark' | 'block' | 'link'
+
+const LIST_TYPES = ['numbered-list', 'bulleted-list']
+
+type CheckIfTypeIsActiveArgs = {
+  type: ButtonType
+  editor: Editor
+  format: string
+}
+
+type ToggleByTypeArgs = {
+  type: ButtonType
+  editor: Editor
+  format: string
+  t: TFunction
+}
+
+type ToggleArgs = Omit<ToggleByTypeArgs, 'type'>
 
 const validateProtocol = (href: string) => {
   const httpStr = 'http://'
@@ -13,17 +34,69 @@ const validateProtocol = (href: string) => {
   return href
 }
 
-export const isLinkActive = (editor: Editor) => {
+const unwrapLink = (editor: Editor) => {
+  Transforms.unwrapNodes(editor, {
+    match: (n) => !Editor.isEditor(n) && SlateElement.isElement(n) && n.type === 'link',
+  })
+}
+
+const isLinkActive = (editor: Editor) => {
   const [link] = Editor.nodes(editor, {
     match: (n) => !Editor.isEditor(n) && SlateElement.isElement(n) && n.type === 'link',
   })
   return !!link
 }
 
-const unwrapLink = (editor: Editor) => {
-  Transforms.unwrapNodes(editor, {
-    match: (n) => !Editor.isEditor(n) && SlateElement.isElement(n) && n.type === 'link',
+const isMarkActive = (editor: Editor, format: string) => {
+  const marks = Editor.marks(editor)
+  return marks ? marks[format] === true : false
+}
+
+const isBlockActive = (editor: Editor, format: string) => {
+  const [match] = Editor.nodes(editor, {
+    match: (n) => n.type === format,
   })
+
+  return !!match
+}
+
+export const isTypeActive = ({ type, editor, format }: CheckIfTypeIsActiveArgs) => {
+  const fn = {
+    mark: isMarkActive,
+    block: isBlockActive,
+    link: isLinkActive,
+  }[type]
+
+  return fn(editor, format)
+}
+
+const toggleBlock = ({ editor, format }: ToggleArgs) => {
+  const isActive = isBlockActive(editor, format)
+  const isList = LIST_TYPES.includes(format)
+
+  Transforms.unwrapNodes(editor, {
+    match: (n) => LIST_TYPES.includes(n.type),
+    split: true,
+  })
+
+  Transforms.setNodes(editor, {
+    type: isActive ? 'paragraph' : isList ? 'list-item' : format,
+  })
+
+  if (!isActive && isList) {
+    const block = { type: format, children: [] }
+    Transforms.wrapNodes(editor, block)
+  }
+}
+
+export const toggleMark = ({ editor, format }: ToggleArgs) => {
+  const isActive = isMarkActive(editor, format)
+
+  if (isActive) {
+    Editor.removeMark(editor, format)
+  } else {
+    Editor.addMark(editor, format, true)
+  }
 }
 
 const wrapLink = (editor: Editor, url: string) => {
@@ -44,6 +117,21 @@ const wrapLink = (editor: Editor, url: string) => {
   } else {
     Transforms.wrapNodes(editor, link, { split: true })
     Transforms.collapse(editor, { edge: 'end' })
+  }
+}
+
+const toggleLink = ({ editor, t }: ToggleArgs) => {
+  let url: string | undefined | null = window.prompt('Enter the URL of the link:')
+  if (!url) return
+
+  url = validateProtocol(url)
+  if (!isUrl(url)) {
+    toast.error(t('websitePatternError'))
+    return
+  }
+
+  if (editor.selection) {
+    wrapLink(editor, url)
   }
 }
 
@@ -75,14 +163,12 @@ export const withLinks = (editor: Editor) => {
   return editor
 }
 
-export const toggleLink = (editor: Editor) => {
-  let url: string | undefined | null = window.prompt('Enter the URL of the link:')
-  if (!url) return
-
-  url = validateProtocol(url)
-  if (!isUrl(url)) return
-
-  if (editor.selection) {
-    wrapLink(editor, url)
+export const toogleByType = ({ type, editor, format, t }: ToggleByTypeArgs) => {
+  const toggles = {
+    mark: toggleMark,
+    block: toggleBlock,
+    link: toggleLink,
   }
+
+  toggles[type]({ editor, format, t })
 }
